requirements.txt
# IVB Bot v2 — Python dependencies
# Install with: pip install -r requirements.txt

# Bloomberg API (must be installed from Bloomberg SDK, not PyPI)
# blpapi — install from C:\blp\DAPI\APIv3\PythonAPI\
# OR via pip: pip install --index-url=https://blpapi.bloomberg.com/repository/releases/python/simple/ blpapi

# Core data & math
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# Machine learning (optional — for ML signal scorer upgrade)
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0

# Terminal dashboard
rich>=13.7.0

# Async & utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0

# API server (optional — for remote monitoring)
fastapi>=0.110.0
uvicorn>=0.27.0
